requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.5
django-allauth==0.42.0
django-appconf==1.0.4
django-crispy-forms==1.9.1
django-formtools==2.2
django-haystack==2.8.1
django-imagekit==4.0.2
django-import-export==2.1.0
emoji==0.5.4
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.17.3
idna==2.8
jdcal==1.4.1
jieba==0.42.1
Markdown==3.2.2
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pilkit==2.0
Pillow==7.1.2
Pygments==2.6.1
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
tablib==2.0.0
urllib3==1.25.9
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
